+++ pkg/agent/as3/as3Common.go
@@ -85,10 +85,8 @@ func (am *AS3Manager) processResourcesForAS3(sharedApp as3Application) {
 //}
 
 func (am *AS3Manager) processIRulesForAS3(sharedApp as3Application) {
-	am.ResourceRequest.IRulesStore.IrulesMutex.Lock()
-	defer am.ResourceRequest.IRulesStore.IrulesMutex.Unlock()
 	// Create irule declaration
-	for _, v := range am.IRulesStore.IRulesMap {
+	for _, v := range am.IrulesMap {
 		iRule := &as3IRules{}
 		iRule.Class = "iRule"
 		iRule.IRule = v.Code